backend/requirements.txt
tensorflow==2.3.0
fastapi==0.60.1
efficientnet==1.1.0
uvicorn==0.11.7
opencv-python==4.3.0.36
Pillow==7.2.0

